Add recognition and skip VS generated visual code
Environment (Visual Studio) generated code for GUI should not be
changed by user to "fix" problematic (rule reported) parts. Changes
in designer files can cause problems in editing of GUI in environment.

Only a small set of problems can be fixed in the GUI editor. One of the
possible fixes is to remove/change used components. (e. g. do not use
the build in web browser, which is not supported for OSX)