Fix incorrect assumptions in duplicate code search
The original assumption was, that only the target method can not be
constructor or generated code, but the source method could be
constructor or generated code. This got too many false positives in my
code, saying that constructor is similar to "Equals" method.

Even the assumption, that constructors can not be similar is wrong.
Two constructor can be similar and (at least some times) they can
be merged using to a single constructor to initialize identical values.