Cool down the duplicate code locator reports
The duplicate code locator is way too aggressive. Unfortunately all
reports are with high severity and normal confidence. This can cause
that even a few similar lines (in worst case scenario only two or three)
are reported as duplicate code (of course the worst case scenario
included generated code, e. g. Linq, dynamic, or other special cases).

The reported severity and confidence was decreased according to
similarity level using "count" that is used in the locator. Therefore
the standard behaviour (first found similarity) was changed to maximal
similarity search.

The duplicate code locator is still way to aggressive, but at least now
it does not seems like this is a totally critical problem.

To do: Fix the locator to do not search in generated (partial) code,
e. g. when using Linq or dynamic in function, and to report only when
the similarity can really be extracted/merged (not when the functions
look similar but have different values in nested calls, or they call
a different function). Similarity can be caused even by the check of
input values (not null, out of range e.t.c.) in public functions.