Fix confidence in unnecessary  specialization rule
The confidence in Avoid Unnecessary Specialization Rule was always
high. But there are reasons when the use of more general type should
not be used. (See example, even when the example is not the best one.)

Therefore the confidence is decreased when function with the same name
has more parameter alternatives (see the new example). Confidence is
also decreased when the only parent is object.

The problem is still reported, but not with the same confidence and
severity. The console version will no longer report problem for the new
example (with standard settings).