Fixed implementation of AvoidUnusedParametersTest
The AnonymousMethodTest implementation was incorrect.
The implementation did use only a type name (without name-space) and
depended on exact generated name. Visual studio generates a different
name. Therefore the test was changed to search for the method in a
correct way. This is not a test to resolve generated method, but a test
for the generated method.