Fix wrong assumptions in framework variable rocks test
The test for Variable Definition Rocks class assumed that there is no
need to test user defined local variables, since they will be present.
When there is no debug information file, all variables are reported
as generated by compiler and no error is raised.

This is IMHO wrong. The test must confirm, that all user variables,
that were defined, are present. Otherwise the test is only a half test.