Update Variable Definition Rocks test
There are two ways how to get used/defined variables. The original test
tested only way. This will add the second way how to get list of local
variables. It also updates the message to determine the real failed
test, not only the method the test failed on. Method name is defined
by test name and it is not necessary to write it once more.