Update 'variable names should not match field' rule
Parameter names should generally not match field names. But (mostly)
for constructors it is hard to reach. It is not the best idea to
'fabricate' a different name for constructor parameter, or field name
only to satisfy this rule. FxCop and Solar Lint do not report identical
names of parameters and field in constructors.

Therefore a simple assignment check was added to the test. Since these
functions should be simple, there should be no problem in most cases.
For now the method do no complain even for a different functions with
direct assignment. After real code test this can be updated to ignore
assignment parameters only for constructors.