Remove MoMA Web Service (Mono Compatibility Download)
The Mono Migration Assistant (MoMA) is in stalled state for a few
years, therefore it has no meaning to download the definition file from
web service. The latest available definition file was added to project
resources. It is still possible to replace the definition file.

To use custom/updated definition file place it to:
${LocalApplicationData}/Gendarme with the name "definitions-2.9.zip"
or newer (version number).