Fix framework 'Is Override' and 'Compare Signature'
The fix for both methods is tested with new 'Is Override' (originally
missing) test. Also added a test for 'Avoid ref and out parameters'
rule. Both method signature changes are a extreme case (not from real
praxis), but are necessary to test special cases. The 'Is Override'
method ignored 'is new slot'