Fix 'Is Main Signature' check function
Since the function 'Is Main' did not check, it it is the application
entry point, function was renamed to do not cause confusion.

In addition, the check for types was updated to do not cause false
positives (User type named Int32 or Void in different name-space).

Also the method 'IsEntryPoint' was removed, since it was used only
in a single rule and even the use was not correct.

(cherry picked from commit 963460c669ba427fc4788b160bcc8306e6ccefb1)